requirements.txt
# Django Core
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Environment variables
python-dotenv>=1.0.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
requests>=2.31.0
tenacity>=8.2.0  # Retry logic

# Caching (optional, for production)
redis>=5.0.0
django-redis>=5.4.0

# CORS
django-cors-headers>=4.3.0

# RAG System
faiss-cpu>=1.7.4
numpy>=1.24.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
pytest-django>=4.5.0

